% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver.R
\name{add_uuid}
\alias{add_uuid}
\title{Add UUID column to make the records unique}
\usage{
add_uuid(df)
}
\arguments{
\item{df}{Data frame. The data frame to modify.}
}
\value{
A data frame with an added UUID column.
}
\description{
Add UUID column to make the records unique
}
\examples{
df <- data.frame(poi_name = c("A", "B"))
df <- add_uuid(df)
}
