% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronze.R
\name{check_approximate_column_names}
\alias{check_approximate_column_names}
\title{Function to check and rename approximate column names to match expected column names}
\usage{
check_approximate_column_names(df)
}
\arguments{
\item{df}{Data frame. The data frame with columns to check and rename.}
}
\value{
A data frame with columns renamed to match the expected column names.
}
\description{
Function to check and rename approximate column names to match expected column names
}
\examples{
df <- data.frame(name = c("Alice", "Bob"), lat = c(34.05, 36.77),
                 lon = c(-118.24, -119.41), connectivity = c(TRUE, FALSE),
                 electricity = c(TRUE, TRUE), poi_subtype = c("school", "hospital"))
df_renamed <- check_approximate_column_names(df)
}
