% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver.R
\name{check_electricity_connectivity}
\alias{check_electricity_connectivity}
\title{Check has_electricity FALSE and is_connected TRUE cases}
\usage{
check_electricity_connectivity(df)
}
\arguments{
\item{df}{Data frame. The data frame to check.}
}
\value{
Data frame with cases where has_electricity is FALSE and is_connected is TRUE.
}
\description{
Check has_electricity FALSE and is_connected TRUE cases
}
\examples{
df <- data.frame(is_connected = c(TRUE, TRUE), has_electricity = c(FALSE, TRUE))
check_electricity_connectivity(df)
}
