% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver.R
\name{check_pois_within_boundaries}
\alias{check_pois_within_boundaries}
\title{Check if POIs are within the country boundary}
\usage{
check_pois_within_boundaries(df, country_boundaries)
}
\arguments{
\item{df}{Data frame. The data frame with POIs.}

\item{country_boundaries}{sf object. The country boundaries.}
}
\value{
Data frame with POIs outside the country boundaries.
}
\description{
Check if POIs are within the country boundary
}
\examples{
df <- data.frame(latitude = c(34.05, -15.78),
                 longitude = c(-118.24, -47.93))
country_boundaries <- sf::st_as_sf(
                          rnaturalearth::ne_countries(
                              scale = "medium",
                              returnclass = "sf"))
check_pois_within_boundaries(df, country_boundaries)
}
