% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronze.R
\name{export_final_df}
\alias{export_final_df}
\title{Function to export the final data frame to a CSV file}
\usage{
export_final_df(df, file_path)
}
\arguments{
\item{df}{Data frame. The data frame to export.}

\item{file_path}{Character. The path to the file where the data frame will be saved.}
}
\description{
Function to export the final data frame to a CSV file
}
