% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier2ott.R
\name{identifier2ott}
\alias{identifier2ott}
\title{Get ott ids for a set of identifiers from four common databases}
\usage{
identifier2ott(
  eol = NA,
  gbif = NA,
  ncbi = NA,
  iucn = NA,
  rate = 5,
  basereq = NA
)
}
\arguments{
\item{eol}{a numeric ID or numeric vector of eol ids indicating the particular taxa to evaluate.}

\item{gbif}{a numeric ID or numeric vector of gbif ids indicating the particular taxa to evaluate.}

\item{ncbi}{a numeric ID or numeric vector of ncbi ids indicating the particular taxa to evaluate.}

\item{iucn}{a numeric ID or numeric vector of iucn ids indicating the particular taxa to evaluate.}

\item{rate}{maximum number of calls to the API per second.}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=oz_basereq]{oz_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A list of maps between the given ids and their OTT counterparts
}
\description{
Retrieve Open Tree of Life IDs for a set of identifiers from four common databases: eol, gbid, ncbi, and iucn.
This is currently \emph{experimental} and does not attempt to do formatting or processing of the JSON output past parsing it into a list.
\strong{Do not use for mission-critical functionality unless you really need to.}
}
\examples{
\dontrun{
identifier2ott(eol = c(1228387, 7674), gbif = c(1651891), ncbi = c(389061), iucn = (7377))
}

}
\author{
Francis Windram
}
