% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_images.R
\name{node_images}
\alias{node_images}
\title{Get get images for a set of ott IDs}
\usage{
node_images(otts, type = "any", rate = 5, key = NA, basereq = NA)
}
\arguments{
\item{otts}{a numeric ID or numeric vector of ott ids indicating the particular taxa to evaluate.}

\item{type}{whether to return any representative images, only images that have been verified as correctly identified, or only public domain images. Defaults to \code{"any"}.}

\item{rate}{maximum number of calls to the API per second.}

\item{key}{your OneZoom API key (defaults to the public key of \code{0} if not specified).}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=oz_basereq]{oz_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A data frame of the output image locations, names, identifiers and their usage rights.
}
\description{
Retrieve reusable, phylogenetically diverse and representative images for a set of Open Tree of Life IDs.
}
\section{Types}{

The following types valid options for the \code{type} argument:
\itemize{
\item \code{"any"} - \emph{any images}
\item \code{"verified"} - \emph{images verified as correctly identified}
\item \code{"pd"} - \emph{images explicitly in the public domain}
}
}

\examples{
\dontrun{
node_images(247341)

node_images(c(247341, 563159, 269666, 504327, 668392), rate=5)

node_images(c(247341, 563159, 269666, 504327, 668392), type="pd")
}

}
\author{
Francis Windram
}
