% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott2common.R
\name{ott2common}
\alias{ott2common}
\title{Get common (vernacular) names for a set of ott IDs}
\usage{
ott2common(
  otts,
  lang = "en",
  findall = FALSE,
  include_unpreferred = FALSE,
  nulls = FALSE,
  rate = 5,
  key = NA,
  basereq = NA
)
}
\arguments{
\item{otts}{a numeric ID or numeric vector of ott ids indicating the particular taxa to evaluate.}

\item{lang}{a language code to specify which language to use.
This can be a 2 or 3 letter code such as \code{"en"} or \code{"fr"}, or a more specific regional code such as \code{"en-gb"} (British English) or \code{"zh-hant"} (Chinese, in traditional script).
Defaults to \code{"en"}}

\item{findall}{return all names for each OTT, in decreasing order of preference.}

\item{include_unpreferred}{if \code{findall=FALSE} and there is no name marked as \emph{"preferred"} for this \code{lang}, allow for return of names marked as \emph{"unpreferred"}.}

\item{nulls}{If no common name is present in the specified \code{lang}, request that the server returns \code{null} instead. Note this is not really necessary to set as this is handled client-side, as such it may be removed.}

\item{rate}{maximum number of calls to the API per second.}

\item{key}{your OneZoom API key (defaults to the public key of \code{0} if not specified).}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=oz_basereq]{oz_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A data frame of the input otts alongside their popularity and popularity rank.
}
\description{
Retrieve common (vernacular) names from onezoom for a set of Open Tree of Life IDs.
}
\section{Column Names}{

If \code{findall=TRUE} is set, the output of this function is arbitrarily wide (as each OTT id has an arbitrary number of common names).
In this case, the columns are numbered after the initial \code{common} column (which is the most preferred name) with names following the formula \code{common_x} where x is the rank of the name between 2 and the largest number of common names returned.
If not all OTTs have the same number of common names, the extra columns will be filled with \code{NA} for those ids.
}

\examples{
\dontrun{
ott2common(247341)

ott2common(c(247341, 563159, 269666, 504327, 668392), rate=5)

ott2common(c(247341, 563159, 269666, 504327, 668392), findall=TRUE)
}

}
\author{
Francis Windram
}
