% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ott2identifiers.R
\name{ott2identifiers}
\alias{ott2identifiers}
\title{Get identifiers from other databases for a set of ott IDs}
\usage{
ott2identifiers(otts, remove_empty = FALSE, rate = 5, key = NA, basereq = NA)
}
\arguments{
\item{otts}{a numeric ID or numeric vector of ott ids indicating the particular taxa to evaluate.}

\item{remove_empty}{remove columns if all entries are \code{NA}.}

\item{rate}{maximum number of calls to the API per second.}

\item{key}{your OneZoom API key (defaults to the public key of \code{0} if not specified).}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=oz_basereq]{oz_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A data frame of the input otts alongside their ids in other databases
}
\description{
Retrieve identifiers from other databases through onezoom for a set of Open Tree of Life IDs.
}
\section{Identifiers}{

The following identifiers are potentially returned
\itemize{
\item \code{ncbi}
\item \code{ifung}
\item \code{worms}
\item \code{irmng}
\item \code{gbif}
\item \code{ipni}
\item \code{eol}
\item \code{wikidata}
\item \code{iucn} (only for leaf node taxa)
}
}

\examples{
\dontrun{
ott2identifiers(247341)

ott2identifiers(c(247341, 563159, 269666, 504327, 668392), rate=5)

ott2identifiers(c(247341, 563159, 269666, 504327, 668392), remove_empty=TRUE)
}

}
\author{
Francis Windram
}
