% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_basereq.R
\name{oz_basereq}
\alias{oz_basereq}
\title{Generate a base request object for the onezoom database}
\usage{
oz_basereq(
  baseurl = "https://www.onezoom.org/",
  useragent = "ROZ",
  unsafe = FALSE
)
}
\arguments{
\item{baseurl}{the base url for the onezoom API.}

\item{useragent}{the user agent string used when contacting onezoom.}

\item{unsafe}{disable ssl verification (should only ever be required on Linux, \strong{do not enable this by default}).}
}
\value{
Returns an httr2 request object, pointing at baseurl using useragent
}
\description{
This request is used as the basis for all calls to the onezoom API.
It does not contain any tokens or session ids, and thus can be regenerated at any time.
}
\examples{
\dontrun{
basereq <- oz_basereq(
  baseurl="https://www.onezoom.org/",
  useragent="ROZ")
}

}
\author{
Francis Windram
}
\concept{basereq}
