% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularity.R
\name{popularity}
\alias{popularity}
\title{Get popularity for a set of ott IDs}
\usage{
popularity(
  otts,
  maxtaxa = NA,
  expand = FALSE,
  spread_evenly = FALSE,
  taxon_names = FALSE,
  include_raw = FALSE,
  sort = NA,
  rate = 5,
  key = NA,
  basereq = NA
)
}
\arguments{
\item{otts}{a numeric ID or numeric vector of ott ids indicating the particular taxa to evaluate.}

\item{maxtaxa}{An integer giving the maximum number of taxa to return.
By default this is equivalent to the number of OTTs passed in: if expand_taxa is \code{TRUE}, you may wish to set this to a higher value.
There is an upper limit on the maximum number of taxa you can ask for, which is determined by your API key.}

\item{expand}{Include all species which are descendants of the given OTTs (i.e. higher level taxa are "expanded" into their constituent species, and the returned taxa are all at the species level).
You can mix OTTs for species and higher taxa in a single call: only higher taxa will be expanded.
If the parameter is \code{FALSE} or not given, do not expand taxa and simply return data on the OTTs that have been passed in (note that higher level taxa will not contain popularity_rank information, which is only valid for species).}

\item{spread_evenly}{Divide the maximum number of taxa to return (e.g. as given by the max parameter) so that each of the N taxa that are passed in are limited to the top (max÷N) most popular species.
Results are then combined and re-sorted, meaning that the final list of taxa returned may well contain fewer species than max.
This only really makes sense when \code{expand=TRUE}.}

\item{taxon_names}{Return a scientific (Latin) name for each returned taxon, if one is available. Names are usually (but not necessarily) the same as the scientific name listed on the Open Tree of Life.}

\item{include_raw}{Also return the raw (non-phylogenetically informed) score for each taxon.}

\item{sort}{If this is set to \code{rank} or \code{raw} then do not order species by phylogenetic popularity score then OTT id (as is the default), but instead use an alternative order.
If \code{raw}, then use the raw (non-phylogenetic) score, which may be zero or absent for some taxa such as these with no wikipedia page (the raw score is also returned, in the same way as it is when setting \code{include_raw=TRUE}).
If \code{rank}, then order by species rank (lowest first, with higher level taxa at start) then by plain popularity score (highest first), and finally by OTT id. Since only species (not higher taxa) have a rank, this necessarily puts all species first.}

\item{rate}{maximum number of calls to the API per second.}

\item{key}{your OneZoom API key (defaults to the public key of \code{0} if not specified).}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=oz_basereq]{oz_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A data frame of the input otts alongside their popularity and popularity rank.
}
\description{
Retrieve popularity indices from onezoom for a set of Open Tree of Life IDs. See \url{https://www.onezoom.org/popularity} for full methodology.
}
\examples{
\dontrun{
popularity(247341)

popularity(c(247341, 563159, 269666, 504327, 668392), rate=5)

popularity(c(247341, 563159, 269666, 504327, 668392), expand=TRUE, spread_evenly=TRUE, rate=5)
}

}
\author{
Francis Windram
}
