% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronze.R
\name{read_example}
\alias{read_example}
\title{Read or list the CSV files}
\usage{
read_example(path = NULL)
}
\arguments{
\item{path}{Character. The name of the file to read from the extdata directory of the roz package. If NULL, lists all example files.}
}
\value{
If the path is NULL, returns a character vector of the names of example files. If the path is specified, returns the full path to the specified example file.
}
\description{
Read or list the CSV files
}
\examples{
read_example()
read_example("raw_data.csv")
}
