% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roz-package.R
\docType{package}
\name{roz-package}
\alias{roz}
\alias{roz-package}
\title{roz: R etl pipeline}
\description{
Does a basic etl pipeline. Convenience functions to make some common tasks with string manipulation and regular expressions a bit easier.
}
\author{
\strong{Maintainer}: Oguzhan Er \email{oguzhanerr@gmail.com}

}
\keyword{internal}
