% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronze.R
\name{transform_to_lower_case}
\alias{transform_to_lower_case}
\title{Function to transform all character data in a data frame to lower case}
\usage{
transform_to_lower_case(df)
}
\arguments{
\item{df}{Data frame. The data frame to transform.}
}
\value{
A data frame with all character data transformed to lower case.
}
\description{
Function to transform all character data in a data frame to lower case
}
\examples{
df <- data.frame(Name = c("Alice", "Bob"), City = c("New York", "Los Angeles"))
df_lower <- transform_to_lower_case(df)
}
