% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver.R
\name{update_electricity}
\alias{update_electricity}
\title{Make has_electricity TRUE if it was NULL and if is_connected is TRUE}
\usage{
update_electricity(df)
}
\arguments{
\item{df}{Data frame. The data frame to modify.}
}
\value{
A data frame with updated has_electricity column.
}
\description{
Make has_electricity TRUE if it was NULL and if is_connected is TRUE
}
\examples{
df <- data.frame(is_connected = c(TRUE, FALSE), has_electricity = c(NA, FALSE))
df <- update_electricity(df)
}
