% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roz-package.R
\docType{package}
\name{roz-package}
\alias{roz}
\alias{roz-package}
\title{roz: Onezoom R Interface}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interface with the public API of the Onezoom project using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fwimp/roz}
  \item \url{https://fwimp.github.io/roz/}
  \item Report bugs at \url{https://github.com/fwimp/roz/issues}
}

}
\author{
\strong{Maintainer}: Francis Windram \email{francis.windram17@imperial.ac.uk} (\href{https://orcid.org/0000-0002-2129-826X}{ORCID})

}
\keyword{internal}
