% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bronze.R
\name{transform_csv_to_df}
\alias{transform_csv_to_df}
\title{Function to read a CSV file, transform it into a data frame, and validate columns and data types}
\usage{
transform_csv_to_df(file_path)
}
\arguments{
\item{file_path}{Character. The path to the CSV file to be read.}
}
\value{
A data frame with only the expected columns if they match, otherwise an error message.
}
\description{
Function to read a CSV file, transform it into a data frame, and validate columns and data types
}
\examples{
file_path <- read_example("raw_data.csv")
df <- transform_csv_to_df(file_path)
}
